Implement Tag List View for Ui
