+++ src/main/java/seedu/address/Main.java
@@ -20,6 +20,6 @@ import javafx.application.Application;
  */
 public class Main {
     public static void main(String[] args) {
-        Application.launch(NewMainApp.class, args);
+        Application.launch(MainApp.class, args);
     }
 }
