+++ src/main/java/seedu/address/MainApp.java
@@ -38,7 +38,7 @@ public class MainApp extends Application {
 
     public static final Version VERSION = new Version(0, 6, 0, true);
 
-    private static final Logger logger = LogsCenter.getLogger(MainApp.class);
+    private static final Logger logger = LogsCenter.getLogger(OldMainApp.class);
 
     protected Ui ui;
     protected Logic logic;
@@ -167,7 +167,7 @@ public class MainApp extends Application {
 
     @Override
     public void start(Stage primaryStage) {
-        logger.info("Starting AddressBook " + MainApp.VERSION);
+        logger.info("Starting AddressBook " + OldMainApp.VERSION);
         ui.start(primaryStage);
     }
