--- src/main/resources/view/NewMainWindow.fxml
@@ -1,68 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.Scene?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.Menu?>
-<?import javafx.scene.control.MenuBar?>
-<?import javafx.scene.control.MenuItem?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.StackPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-<?import javafx.stage.Stage?>
-
-<fx:root onCloseRequest="#handleExit" title="HelloFile App" type="Stage" xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1">
-   <scene>
-      <Scene>
-         <root>
-            <VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="501.0">
-               <children>
-                  <MenuBar>
-                    <menus>
-                      <Menu mnemonicParsing="false" text="File">
-                        <items>
-                          <MenuItem mnemonicParsing="false" text="Exit" />
-                        </items>
-                      </Menu>
-                      <Menu mnemonicParsing="false" text="Help">
-                        <items>
-                          <MenuItem fx:id="helpMenuItem" mnemonicParsing="false" text="Help" />
-                        </items>
-                      </Menu>
-                    </menus>
-                  </MenuBar>
-                  <HBox prefHeight="100.0" prefWidth="200.0" VBox.vgrow="ALWAYS">
-                     <children>
-                        <VBox prefHeight="314.0" prefWidth="188.0" HBox.hgrow="NEVER">
-                           <children>
-                              <Label text="Tags:" VBox.vgrow="NEVER">
-                                 <font>
-                                    <Font size="16.0" />
-                                 </font>
-                              </Label>
-                              <StackPane fx:id="tagListPlaceholder" prefHeight="306.0" prefWidth="139.0" VBox.vgrow="ALWAYS" />
-                           </children>
-                        </VBox>
-                        <VBox prefHeight="200.0" prefWidth="100.0" spacing="10.0" HBox.hgrow="ALWAYS">
-                           <children>
-                              <StackPane fx:id="resultDisplayPlaceHolder" prefHeight="299.0" prefWidth="359.0" VBox.vgrow="ALWAYS" />
-                              <StackPane fx:id="lastInputPlaceHolder" prefHeight="15.0" prefWidth="301.0" VBox.vgrow="NEVER" />
-                           </children>
-                        </VBox>
-                     </children>
-                     <padding>
-                        <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
-                     </padding>
-                  </HBox>
-                  <StackPane fx:id="commandBoxPlaceHolder" prefHeight="41.0" prefWidth="501.0">
-                     <padding>
-                        <Insets left="10.0" right="10.0" />
-                     </padding></StackPane>
-                  <StackPane fx:id="footerbarPlaceHolder" prefHeight="19.0" prefWidth="501.0" />
-               </children>
-            </VBox>
-         </root>
-      </Scene>
-   </scene>
-</fx:root>
